backend/requirements.txt
fastapi
uvicorn
boto3
python-multipart
python-dotenv
pillow
langchain
langgraph
langchain-aws
langchain-community
ruff
pytest
